requirements.txt
chia-blockchain==1.2.11